fixing the analysis code

+++ common/include/dfd_net_v14.h
@@ -141,7 +141,7 @@ using dfd_net_type = dlib::loss_multiclass_log_per_pixel<
     dtago2<dfd_res_33<512, 512, con2d<512,
     
     dtago1<dfd_res_33<256, 256, cbp3_blk<256, 
-    dlib::tag10<dlib::input_dfd_array<uint16_t, img_depth>>
+    dlib::tag10<dlib::input_dfd_array<uint8_t, img_depth>>
     //dlib::input_dfd_array<uint16_t, img_depth>
     >>> >>> >>>> > >>>> > >>>>> >;
     
@@ -161,7 +161,7 @@ using adfd_net_type = dlib::loss_multiclass_log_per_pixel<
     
     dtago1<adfd_res_33<256, 256, acbp3_blk<256, 
     //dlib::input<std::array<dlib::matrix<uint16_t>, img_depth>
-    dlib::tag10<dlib::input_dfd_array<uint16_t, img_depth>>
+    dlib::tag10<dlib::input_dfd_array<uint8_t, img_depth>>
     >>> >>> >>>> > >>>> > >>>>> >;
