+++ common/include/dfd_net_v15b.h
@@ -139,7 +139,7 @@ using dfd_net_type = dlib::loss_multiclass_log_per_pixel<
     dtago2<dfd_res_33<512, 512, con2d<512,
     
     dtago1<dfd_res_55<256, 256, cbp5_blk<256, 
-    dlib::tag10<dlib::input_dfd_array<uint16_t, img_depth>>
+    dlib::tag10<dlib::input_dfd_array<uint8_t, img_depth>>
     //dlib::input_dfd_array<uint16_t, img_depth>
     >>> >>> >>>> > >>>> > >>>>> >;
     
@@ -159,7 +159,7 @@ using adfd_net_type = dlib::loss_multiclass_log_per_pixel<
     
     dtago1<adfd_res_55<256, 256, acbp5_blk<256, 
     //dlib::input<std::array<dlib::matrix<uint16_t>, img_depth>
-    dlib::tag10<dlib::input_dfd_array<uint16_t, img_depth>>
+    dlib::tag10<dlib::input_dfd_array<uint8_t, img_depth>>
     >>> >>> >>>> > >>>> > >>>>> >;
